file cleanup
moved validates.php to the includes folder, renamed index to server_lookup, and deleted redundant toystore files like toy.php
